feat: Use health-check endpoints
Let the Docker Compose files use health-check endpoints to determine
whether a service is healthy and ready to accept connections.

The development Compose file now also waits for the "frontend" service
to be healthy before starting the "app" service, ensuring that the
application server (with its Caddy reverse proxy) only starts after the
frontend is ready.