chore: Change the mountpoints for the "data" and "caddy" volumes
The "caddy" volume is now mounted at "/opt/bacmet/.local/share/caddy".
With the container-user's home directorf buing "/opt/bacmet", this is
the default location for Caddy to store its data, and no additional
configuration is needed for Caddy to use this location.

The "data" volume is now mounted at "/opt/bacmet/data".  This is a
more intuitive location for the application's data, and it keeps all
application-related files under the "/opt/bacmet" directory.